Developed the GFF-table further. Added extern.R. Replaced cmsearch16S with barrnap.

+++ R/blasting.R
@@ -81,30 +81,31 @@
 #' @importFrom microseq gregexpr
 #' @export
 blastAllAll <- function( in.files, out.folder, e.value=1, job=1, threads=1, verbose=T ){
-  for( i in 1:length( in.files ) ){
-    command <- paste( "makeblastdb -logfile log.txt -dbtype prot -out blastDB", job, " -in ", in.files[i], sep="" )
-    system( command )
-    file.remove( "log.txt" )
-    gi <- gregexpr( "GID[0-9]+", in.files[i], extract=T )
-    for( j in 1:length( in.files ) ){
-      gj <- gregexpr( "GID[0-9]+", in.files[j], extract=T )    
-      rname <- paste( gj, "_vs_", gi, ".txt", sep="" )
-      res.files <- dir( out.folder )
-      if( !(rname %in% res.files) ){
-        if( verbose ) cat( "blastAllAll: ", rname, "\n" )
-        input <- paste( "-query ", in.files[j], sep="" )
-        dbase <- paste( "-db blastDB", job, sep="" )
-        output <- paste( "-out ", file.path( out.folder, rname ), sep="" )
-        command <- paste( "blastp -matrix BLOSUM45 -evalue", e.value, "-num_threads", threads,
-                          "-outfmt 6", input, dbase, output )
-        system( command )
+  if( available.external( "blast+" ) ){
+    for( i in 1:length( in.files ) ){
+      command <- paste( "makeblastdb -logfile log.txt -dbtype prot -out blastDB", job, " -in ", in.files[i], sep="" )
+      system( command )
+      gi <- gregexpr( "GID[0-9]+", in.files[i], extract=T )
+      for( j in 1:length( in.files ) ){
+        gj <- gregexpr( "GID[0-9]+", in.files[j], extract=T )    
+        rname <- paste( gj, "_vs_", gi, ".txt", sep="" )
+        res.files <- dir( out.folder )
+        if( !(rname %in% res.files) ){
+          if( verbose ) cat( "blastAllAll: ", rname, "\n" )
+          input <- paste( "-query ", in.files[j], sep="" )
+          dbase <- paste( "-db blastDB", job, sep="" )
+          output <- paste( "-out ", file.path( out.folder, rname ), sep="" )
+          command <- paste( "blastp -matrix BLOSUM45 -evalue", e.value, "-num_threads", threads,
+                            "-outfmt 6", input, dbase, output )
+          system( command )
+        }
       }
     }
+    file.remove( paste( "blastDB", job, ".pin", sep="" ) )
+    file.remove( paste( "blastDB", job, ".phr", sep="" ) )
+    file.remove( paste( "blastDB", job, ".psq", sep="" ) )
+    return( "done" )
   }
-  file.remove( paste( "blastDB", job, ".pin", sep="" ) )
-  file.remove( paste( "blastDB", job, ".phr", sep="" ) )
-  file.remove( paste( "blastDB", job, ".psq", sep="" ) )
-  return( "done" )
 }
